feat: Reorganize tests into package structure
- Create tests/ subdirectories in research and users apps.

- Split monolithic tests.py files into component-specific files (models, forms, views, utils).

- Add __init__.py files to make directories Python packages.

- Move relevant test classes and methods to new files.

- Delete old tests.py files.
